+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 
 ### Improvements
 
+- [beta][Big Sur] Match the inspector background to desktop color.
 - Round the corners of current line highlight.
 - [beta][tirivial] Update titlebar color.